requirements.txt
attrs>=18.2.0
Flask>=1.0.2
Flask-Admin>=1.5.2
Flask-SQLAlchemy>=2.3.2
Flask-Migrate>=2.3.0
PyYAML>=3.13
